# docs/api.yaml
openapi: 3.0.0
info:
  title: JVN Chatbot API
  version: 1.0.0
  description: |
    Johann von Neumann Chatbot API - A gamified chatbot system where users earn tokens by inspiring JVN through various activities, then spend those tokens to ask questions and receive mathematical wisdom.
    
    ## Core Concepts
    
    **Token Economy**: Users earn tokens (1-10 per activity) by successfully inspiring JVN. One token = 5 questions.
    
    **Mood System**: JVN's mood affects his responses and becomes pickier over time (goes from pleased to analytical to demanding to insufferable).
    
    **Birthday Celebrations**: Every 1000 interactions triggers a birthday mode with special rewards and easier token earning.
    
    **Philosophy Engine**: JVN occasionally shares mathematical wisdom based on his current mood and interaction count.

  contact:
    name: API Support
    url: https://github.com/your-username/jvn-chatbot-api
    email: a2z.tom@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.jvn-chatbot.com
    description: Production server
...
paths:
...
  /chat/ask-question:
    post:
      tags:
        - chat
      summary: Ask JVN a question using tokens
      description: |
        Submit questions to JVN using earned tokens. Each token provides 5 questions.
        
        **Token Usage**: 1 token = 5 questions
        
        JVN's response style varies based on his current mood and may include bonus philosophical moments.
      operationId: askQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  description: The question to ask JVN
                  minLength: 1
                  example: "What is the meaning of mathematics?"
            examples:
              math_question:
                summary: Mathematical question
                value:
                  question: "What makes mathematics beautiful?"
              philosophy_question:
                summary: Philosophical question
                value:
                  question: "How does logic guide decision making?"
      responses:
        '200':
          description: JVN's response to the question
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Success"
                  data:
                    $ref: '#/components/schemas/QuestionResponse'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '400':
          description: Invalid question or no tokens available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                no_tokens:
                  summary: No tokens available
                  value:
                    success: false
                    error:
                      message: "No tokens available! Please inspire JVN to earn tokens first."
                      code: "NO_TOKENS_AVAILABLE"
                      global_token_pool: 0
                      suggestion: "Try POST /inspire/fresh-air or other inspiration endpoints!"
                    meta:
                      timestamp: "2025-01-15T10:30:00.000Z"
                missing_question:
                  summary: Missing question parameter
                  value:
                    success: false
                    error:
                      message: "Ze question parameter is required for mathematical discourse!"
                      code: "MISSING_QUESTION"
                    meta:
                      timestamp: "2025-01-15T10:30:00.000Z"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    QuestionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        jvn_response:
          type: string
          description: JVN's response to the question
          example: "Ah, an interesting question! Let me apply my mathematical reasoning..."
        your_question:
          type: string
          description: The question that was asked
          example: "What is the meaning of mathematics?"
        jvn_mood:
          type: string
          enum: [pleased, analytical, slightly_annoyed, demanding, insufferable, ecstatic]
          description: JVN's mood when answering
          example: "pleased"
        questions_remaining:
          type: integer
          description: Questions remaining from current token
          example: 4
        global_token_pool:
          type: integer
          description: Total tokens available
          example: 12
        total_questions_asked:
          type: integer
          description: Total questions asked in conversation history
          example: 128
        bonus_wisdom:
          type: boolean
          description: Whether bonus philosophical wisdom was included
        jvn_wisdom:
          type: string
          description: Bonus philosophical wisdom if triggered
          example: "Every equation tells a story. Ze art is learning to read ze narrative hidden in ze symbols."
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              description: Human-readable error message
            code:
              type: string
              description: Machine-readable error code
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of response
          example: "2025-01-15T10:30:00.000Z"
  responses:
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: "Ach! A computational catastrophe occurred!"
              code: "INTERNAL_SERVER_ERROR"
            meta:
              timestamp: "2025-01-15T10:30:00.000Z"
...